requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==42.0.3
defusedxml==0.7.1
Django==4.2.7
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-recaptcha==4.0.0
django-rosetta==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
polib==1.2.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.0
Werkzeug==3.0.1
